Fix and refactor security configuration for 1.5
- need to whitelist /js and /css directories
- use l10n.security.authenticationType instead of independant setting for header and oauth
- split in mutliple files
