+++ pom.xml
@@ -12,8 +12,7 @@
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-<!--        <version>1.4.7.RELEASE</version>-->
-        <version>1.5.21.RELEASE</version>
+        <version>1.5.22.RELEASE</version>
     </parent>
 
     <properties>
