+++ webapp/src/main/java/com/box/l10n/mojito/security/SecurityConfig.java
@@ -1,9 +1,10 @@
 package com.box.l10n.mojito.security;
 
-import java.util.Arrays;
-import java.util.List;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
+
+import java.util.Arrays;
+import java.util.List;
 
 /**
  * Contains the security configuration.
@@ -31,6 +32,8 @@ public class SecurityConfig {
         LDAP,
         DATABASE,
         AD,
+        HEADER,
+        OAUTH2
     }
 
 }
