+++ webapp/src/main/java/com/box/l10n/mojito/security/WebSecurityConfig.java
@@ -1,15 +1,11 @@
 package com.box.l10n.mojito.security;
 
+import com.google.common.base.Preconditions;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.boot.autoconfigure.security.Http401AuthenticationEntryPoint;
-import org.springframework.boot.autoconfigure.security.oauth2.resource.ResourceServerProperties;
-import org.springframework.boot.autoconfigure.security.oauth2.resource.UserInfoTokenServices;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.web.servlet.FilterRegistrationBean;
 import org.springframework.context.annotation.AdviceMode;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -20,13 +16,7 @@ import org.springframework.security.config.annotation.method.configuration.Enabl
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.core.userdetails.UserDetailsByNameServiceWrapper;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.oauth2.client.OAuth2ClientContext;
-import org.springframework.security.oauth2.client.OAuth2RestTemplate;
-import org.springframework.security.oauth2.client.filter.OAuth2ClientAuthenticationProcessingFilter;
-import org.springframework.security.oauth2.client.filter.OAuth2ClientContextFilter;
-import org.springframework.security.oauth2.client.token.grant.code.AuthorizationCodeResourceDetails;
 import org.springframework.security.web.authentication.LoginUrlAuthenticationEntryPoint;
 import org.springframework.security.web.authentication.preauth.PreAuthenticatedAuthenticationProvider;
 import org.springframework.security.web.authentication.preauth.RequestHeaderAuthenticationFilter;
@@ -60,18 +50,18 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
     @Autowired
     UserDetailsContextMapperImpl userDetailsContextMapperImpl;
 
-    @Autowired
-    OAuth2ClientContext oauth2ClientContext;
+    @Autowired(required = false)
+    @Qualifier("oauth2Filter")
+    Filter oauth2Filter;
 
-    @Value("${l10n.security.oauth2.enabled:false}")
-    boolean oauth2Enabled = false;
+    @Autowired(required = false)
+    RequestHeaderAuthenticationFilter requestHeaderAuthenticationFilter;
 
-    @Value("${l10n.security.header-auth.enabled:false}")
-    boolean headerAuth = false;
+    @Autowired(required = false)
+    PreAuthenticatedAuthenticationProvider preAuthenticatedAuthenticationProvider;
 
     @Autowired
     public void configureGlobal(AuthenticationManagerBuilder auth) throws Exception {
-
         for (SecurityConfig.AuthenticationType authenticationType : securityConfig.getAuthenticationType()) {
             switch (authenticationType) {
                 case DATABASE:
@@ -82,6 +72,9 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
                     break;
                 case AD:
                     configureActiveDirectory(auth);
+                    break;
+                case HEADER:
+                    configureHeaderAuth(auth);
                     break;
             }
         }
@@ -127,31 +120,10 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
                 .ldif(ldapConfig.getLdif());
     }
 
-    @Autowired
-    void configureHeaderAuth(AuthenticationManagerBuilder auth) throws Exception {
-        if (headerAuth) {
-            logger.debug("Configuring in pre authentication");
-            auth.authenticationProvider(preAuthenticatedAuthenticationProvider());
-        }
-    }
-
-    @Bean
-    @ConditionalOnProperty(value = "l10n.security.header-auth.enabled", havingValue = "true")
-    RequestHeaderAuthenticationFilter requestHeaderAuthenticationFilter() throws Exception {
-        RequestHeaderAuthenticationFilter requestHeaderAuthenticationFilter = new RequestHeaderAuthenticationFilter();
-        requestHeaderAuthenticationFilter.setPrincipalRequestHeader("x-forwarded-user");
-        requestHeaderAuthenticationFilter.setExceptionIfHeaderMissing(false);
-        requestHeaderAuthenticationFilter.setAuthenticationManager(authenticationManager());
-        return requestHeaderAuthenticationFilter;
-    }
-
-    @Bean
-    @ConditionalOnProperty(value = "l10n.security.header-auth.enabled", havingValue = "true")
-    PreAuthenticatedAuthenticationProvider preAuthenticatedAuthenticationProvider() {
-        PreAuthenticatedAuthenticationProvider preAuthenticatedAuthenticationProvider = new PreAuthenticatedAuthenticationProvider();
-        UserDetailsByNameServiceWrapper userDetailsByNameServiceWrapper = new UserDetailsByNameServiceWrapper(getUserDetailsServiceCreatePartial());
-        preAuthenticatedAuthenticationProvider.setPreAuthenticatedUserDetailsService(userDetailsByNameServiceWrapper);
-        return preAuthenticatedAuthenticationProvider;
+    void configureHeaderAuth(AuthenticationManagerBuilder auth) {
+        Preconditions.checkNotNull(preAuthenticatedAuthenticationProvider, "The preAuthenticatedAuthenticationProvider must be configured");
+        logger.debug("Configuring in pre authentication");
+        auth.authenticationProvider(preAuthenticatedAuthenticationProvider);
     }
 
     @Override
@@ -162,8 +134,9 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
 
         http.csrf().ignoringAntMatchers("/shutdown", "/api/rotation");
 
+
         http.authorizeRequests()
-                .antMatchers("/intl/*", "/img/*", "/fonts/*", "/login/**", "/webjars/**", "/cli/**", "/health").permitAll()
+                .antMatchers("/intl/*", "/img/*", "/fonts/*", "/login/**", "/webjars/**", "/cli/**", "/health", "/js/**", "/css/**").permitAll()
                 .antMatchers("/shutdown", "/api/rotation").hasIpAddress("127.0.0.1").anyRequest().permitAll()
                 .anyRequest().fullyAuthenticated()
                 .and()
@@ -173,58 +146,26 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
                 .and()
                 .logout().logoutSuccessUrl("/login?logout").permitAll();
 
-        if (headerAuth) {
-            http.addFilterBefore(requestHeaderAuthenticationFilter(), BasicAuthenticationFilter.class);
-        }
 
-        if (oauth2Enabled) {
-            http.addFilterBefore(oauthFilter(), BasicAuthenticationFilter.class);
-        }
-
+        http.exceptionHandling().defaultAuthenticationEntryPointFor(new LoginUrlAuthenticationEntryPoint("/login"), new AntPathRequestMatcher("/*"));
         http.exceptionHandling().defaultAuthenticationEntryPointFor(new Http401AuthenticationEntryPoint("API_UNAUTHORIZED"), new AntPathRequestMatcher("/api/*"));
-        http.exceptionHandling().defaultAuthenticationEntryPointFor(new LoginUrlAuthenticationEntryPoint(oauth2Enabled ? "/login/oauth" : "/login"), new AntPathRequestMatcher("/*"));
-    }
 
-    private Filter oauthFilter() {
-        logger.debug("Setup SSO filter for oauth");
-        OAuth2ClientAuthenticationProcessingFilter oauth2Filter = new OAuth2ClientAuthenticationProcessingFilter(
-                "/login/oauth");
-        OAuth2RestTemplate auth2RestTemplate = new OAuth2RestTemplate(oauth2(), oauth2ClientContext);
-
-        oauth2Filter.setRestTemplate(auth2RestTemplate);
-        UserInfoTokenServices tokenServices = new UserInfoTokenServices(
-                oauth2Resource().getUserInfoUri(),
-                oauth2().getClientId());
-        tokenServices.setRestTemplate(auth2RestTemplate);
-
-        oauth2Filter.setTokenServices(new MyUserInfoTokenServices(
-                oauth2Resource().getUserInfoUri(),
-                oauth2().getClientId()));
-
-        return oauth2Filter;
-    }
-
-    @Bean
-    @ConditionalOnProperty(value = "l10n.security.oauth2.enabled", havingValue = "true")
-    @ConfigurationProperties("l10n.security.oauth2.client")
-    public AuthorizationCodeResourceDetails oauth2() {
-        return new AuthorizationCodeResourceDetails();
-    }
-
-    @Bean
-    @ConditionalOnProperty(value = "l10n.security.oauth2.enabled", havingValue = "true")
-    @ConfigurationProperties("l10n.security.oauth2.resource")
-    public ResourceServerProperties oauth2Resource() {
-        return new ResourceServerProperties();
-    }
-
-    @Bean
-    @ConditionalOnProperty(value = "l10n.security.oauth2.enabled", havingValue = "true")
-    public FilterRegistrationBean oauth2ClientFilterRegistration(OAuth2ClientContextFilter filter) {
-        FilterRegistrationBean registration = new FilterRegistrationBean();
-        registration.setFilter(filter);
-        registration.setOrder(-100);
-        return registration;
+        for (SecurityConfig.AuthenticationType authenticationType : securityConfig.getAuthenticationType()) {
+            switch (authenticationType) {
+                case HEADER:
+                    Preconditions.checkNotNull(requestHeaderAuthenticationFilter, "The requestHeaderAuthenticationFilter must be configured");
+                    logger.debug("Add request header Auth filter");
+                    requestHeaderAuthenticationFilter.setAuthenticationManager(authenticationManager());
+                    http.addFilterBefore(requestHeaderAuthenticationFilter, BasicAuthenticationFilter.class);
+                    break;
+                case OAUTH2:
+                    Preconditions.checkNotNull(oauth2Filter, "The OAuth2 filter must be configured");
+                    logger.debug("Add OAuth2 filter");
+                    http.addFilterBefore(oauth2Filter, BasicAuthenticationFilter.class);
+                    http.exceptionHandling().defaultAuthenticationEntryPointFor(new LoginUrlAuthenticationEntryPoint("/login/oauth"), new AntPathRequestMatcher("/*"));
+                    break;
+            }
+        }
     }
 
     @Primary
@@ -233,8 +174,4 @@ public class WebSecurityConfig extends WebSecurityConfigurerAdapter {
         return new UserDetailsServiceImpl();
     }
 
-    @Bean
-    protected UserDetailsServiceCreatePartialImpl getUserDetailsServiceCreatePartial() {
-        return new UserDetailsServiceCreatePartialImpl();
-    }
 }
